fix: change input output shape and move UTs
